requirements.txt
# 核心依赖
tos>=2.6.0
volcengine-python-sdk>=1.0.0
httpx>=0.24.0
requests>=2.28.0
pyyaml>=6.0
Pillow>=9.0.0
moviepy>=1.0.3
numpy>=1.21.0
pydantic>=2.0.0

# 工具库
python-dotenv>=1.0.0
aiohttp>=3.8.0
tqdm>=4.64.0

# 日志
loguru>=0.7.0

# 数据库
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
redis>=5.0.0
